% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCsim.R
\name{labels.dc}
\alias{labels.dc}
\title{Get the variable labels of a draws component for a vector-valued parameter}
\usage{
\method{labels}{dc}(object, ...)
}
\arguments{
\item{object}{a component of a draws object.}

\item{...}{currently not used.}
}
\value{
The variable labels.
}
\description{
Get the variable labels of a draws component for a vector-valued parameter
}
\examples{
\donttest{
ex <- mcmcsae_example()
sampler <- create_sampler(ex$model, data=ex$dat)
sim <- MCMCsim(sampler, store.all=TRUE)
labels(sim$beta)
labels(sim$v)
}

}
