% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcsae.R
\docType{package}
\name{mcmcsae-package}
\alias{mcmcsae-package}
\alias{mcmcsae}
\title{MCMC Small Area Estimation}
\description{
Fit multi-level models with possibly correlated random effects using MCMC.
}
\details{
Functions to fit multi-level models with Gaussian and binomial or negative
binomial likelihoods using MCMC. Models with a linear predictor consisting
of various possibly correlated random effects are supported, allowing
flexible modeling of temporal, spatial or other kinds of dependence structures.
For Gaussian models the variance can be modeled too. By modeling unit-level
variances it is possible to account for outliers.
The package has been developed with applications to small area estimation
in official statistics in mind. The posterior samples for the model
parameters can be passed to a prediction function to generate samples from
the posterior predictive distribution for user-defined quantities such as
finite population domain means. For model assessment, posterior predictive
checks and DIC/WAIC criteria are supported.
}
