% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{pr_fixed}
\alias{pr_fixed}
\title{Create an object containing information about a degenerate prior fixing a
parameter to a fixed value}
\usage{
pr_fixed(value = 1, n = NULL, post = FALSE)
}
\arguments{
\item{value}{value parameter of length 1 or \code{n}.}

\item{n}{dimension, if known. For internal use only.}

\item{post}{whether conditional posterior sampling function should be
created. For internal use only.}
}
\value{
An environment with information about the prior and possibly conditional
 posterior distribution(s), to be used by other package functions.
}
\description{
Create an object containing information about a degenerate prior fixing a
parameter to a fixed value
}
