% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{pr_gig}
\alias{pr_gig}
\title{Create an object containing information about Generalized Inverse Gaussian (GIG) prior distributions}
\usage{
pr_gig(a, b, p, n = NULL, post = FALSE)
}
\arguments{
\item{a}{parameter of length 1 or \code{n}.}

\item{b}{parameter of length 1 or \code{n}.}

\item{p}{parameter of length 1 or \code{n}.}

\item{n}{dimension, if known. For internal use only.}

\item{post}{whether conditional posterior sampling function should be
created. For internal use only.}
}
\value{
An environment with information about the prior and possibly conditional
 posterior distribution(s), to be used by other package functions.
}
\description{
Create an object containing information about Generalized Inverse Gaussian (GIG) prior distributions
}
